+++ trunk/MAC/APL/PAC/RSPDriver/src/SetWeightsCmd.cc
@@ -43,7 +43,7 @@ SetWeightsCmd::SetWeightsCmd(RSPSetweightsEvent& sw_event, GCFPortInterface& por
   RSPSetweightsEvent* event = new RSPSetweightsEvent();
   m_event = event;
   
-  event->timestamp = sw_event.timestamp + timestep;
+  event->timestamp = sw_event.timestamp + (long)timestep;
   event->rcumask   = sw_event.rcumask;
 
   setOperation(oper);
